common-utils-py312/requirements.txt
pymongo   
certifi
PyJWT
python-dotenv
cryptography
uuid6
bcrypt